moves axios to the new utils folder allows now post and put to pass parameters too adds account as an possible default value in the baseURL implements the withFullResponse param to get the full axios response and not just the content
